+++ js/browser/wa-crsqlite/package.json
@@ -7,11 +7,11 @@
     "src"
   ],
   "exports": {
-    ".": "./dist/wrapper.js",
+    ".": "./dist/index.js",
     "./wa-sqlite-async.wasm": "./dist/wa-sqlite-async.wasm"
   },
-  "module": "dist/wrapper.js",
-  "types": "dist/wrapper.d.ts",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
   "sideEffects": false,
   "devDependencies": {
     "typescript": "^4.8.4",
