+++ js/browser/wa-crsqlite/src/log.ts
@@ -1,0 +1,6 @@
+const isDebug = (globalThis as any).__vlcn_wa_crsqlite_dbg;
+export default function log(...data: any[]) {
+  if (isDebug) {
+    console.log("wa-crsqlite: ", ...data);
+  }
+}
