+++ js/tests/browser-tests/cypress/component/wa-sqlite-wrapper.cy.ts
@@ -46,3 +46,6 @@ describe("wa-sqlite-wrapper.cy.ts", () => {
     ]);
   });
 });
+
+// TODO:
+// test transaction coordination
